Refactored SVG data tracking.
Instead of simply storing the d value of the SVG path the path was swapped for a polyline and now the entire SVG contents are stored in the hidden field. This puts us one step closer to being able to add swatches so you can doodle / sign in different colors.