Fixed some rendering bugs.
Fill color was defaulting to black on WebKit (or is it AppleWebKit?) based browsers. Updated tracking to allow the mouse to leave the canvas and it still track correctly. Will need to update the code to not record the coordinates if they are outside of the canvas (as to not cause bloat in the generated XML). Already removed the extra spaces that were showing up in the XML.