Refactored the draw recording.
Instead of recording every single path, there is now tracking to see if the path lives inside the canvas element. If it does, the path is recorded to the hidden input, else it's tossed. This refactoring also fixed the issue with the single click dots not showing up on the rendered SVG.